backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Authentication
pyjwt==2.8.0

# Database
psycopg2-binary==2.9.11
asyncpg==0.30.0

# Document Processing
pypdf==5.1.0
python-docx==1.1.2
pdfplumber==0.11.8
beautifulsoup4==4.12.3

# OCR & Image Processing
pytesseract==0.3.10
pillow==10.1.0
numpy==1.24.3
# paddlepaddle==3.2.2  # 依赖过于复杂，暂时禁用
# paddleocr==2.7.0.3

# AI & ML
openai==1.62.0
instructor==1.8.0

# Data Validation
pydantic==2.10.0
pydantic-settings==2.12.0
email-validator==2.1.0

# Logging
loguru==0.7.3

# Cache & Queue
redis==5.2.0
celery[redis]==5.4.0

# Utilities
python-dotenv==1.0.0
aiofiles==24.1.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.0
pytest-cov==4.1.0
